Remove sass dependency
Given that Ruby sass is now deprecated and developers are suggested to
move to sassc, we should stop specifying sass as a runtime dependency.

Also, since it seems like we required sass only just to add to
load_paths, it should be safe to silently fail if it is not available.